Checkpoint of SMS/MMS Recycler.
The recycler allows a user to turn on setting that'll limit the number
of messages saved per thread. An eventual UI will allow users to lock
messages and prevent them from being deleted. This change adds a couple
of tweaks to be able to query the number of messages per thread id, etc.