Fix has_attachment triggers
If you create a thread by sending an MMS text message to an email
address and then create another thread by sending an MMS with an
attachment to a different address, both threads would show an
attachment icon in the conversation list. In the MMS provider, there
are four triggers that deal with setting the thread's has_attachment
column. Two of the triggers were wrong and needed an extra comparison
to only select parts that matched the target thread id. Bug 2235425.

Change-Id: I497a3243f56de216049f18eaa0e96898679f7c62