Fix thread not read
One of the triggers for setting a thread's "read" status wasn't narrowing
its search for message types that can be read. As a result, an unread
"deliver ind" in the pdu that isn't displayed to the user caused the
thread table's "read" column to be set to 0. The other triggers, such as
message_count, only count the pdu messages the user can see. Update this
particular trigger the same way. Bug 2511065

Change-Id: I259717f9692fe68fd15ab7cb1583bc109e93e5f7