Delete unreferenced phone numbers
Bug 5375782. Bug 5172594
The canonical_addresses table in mmssms.db is used to keep a reference
between recipientIds and phone numbers/email addresses. The threads
table contains references to these recipient ids. When a thread is
deleted, the recipients are orphaned. With this change, whenever a
thread is deleted, we delete unreferenced numbers.

Change-Id: I7e1b7a11317d2e6662fab3f434239fe21ca313ec