Message received with wrong sender
Bug 5622048
If the last message of a thread is deleted which results in the very last
recipient in the canonical_addresses table getting deleted, and the messaging
app is still running, and a new message arrives for an address the messaging
app has never seen before, then the new message will appear to have been
sent from the recently deleted recipient. The recipient id cache in the
app doesn't know the recipient was deleted and its cache it out of date.
The fix is to never reuse recipient ids by setting the _id on the
canonical_addresses table to AUTOINCREMENT.

Change-Id: Ia3618fb906e09b3a305ac778b8d73983586fa938