Deleting a thread will mess up a multi-recipient thread with the same recipient
Bug 5763490
The previous DB query for deleting unreferenced recipient id's didn't work
on threads addressed to multiple recipients because the recipient id's are
stored as a space-separated list of ids in a single column. With this change,
the work for identifying unreferenced recipient id's is done in java,
rather than in an sqlite3 selection statement.

Change-Id: Id6ad32408a67dc7550b32ce8ab095635c324a799