Old thumbnail shown
Bug 6602473

Don't reuse part and pdu ids because their uri's are used as keys for the
pdu and thumbnail caches. Change the part and pdu tables to use
AUTOINCREMENT keys. When copying over a table to a temp table, dropping the
original, and renaming the temp to the original, all triggers associated
with the original table are dropped. Those triggers have to be recreated.
The changes look more extensive than they really are. There were many mms-related
triggers in createCommonTriggers() that had to be moved to CreateMmsTriggers().
I used a diff tool to compare the schema dumps from the database before
and after to verify the tables and triggers are the same (and used
that same tool in the first place to identify those missing triggers).

Change-Id: I92a100db7479e28f3c1d62ff6adab961c5871113