Fix permission check for reading SMS on UICC.
When accessing SMS stored on UICC via TelephonyProvider, we need to
use the com.android.phone UID for the call to IccSmsInterfaceManager
to avoid a UID mismatch failure in the app ops check. Add code to
SmsProvider to save and restore the original caller identity.
Also set android:multiprocess attribute to false in the manifest so
that the providers don't run inside the caller's process, which
also leads to a security exception.

Bug: 10131077
Change-Id: Ie6d3477b14973632f3706201de91353950359ee0