36367212, 36366196 Restore MMS attachments from backup file
* Update the TelephonyProvider's backup/restore agent to handle
restoring mms attachments. This change also includes restoring the
original SMIL, and backing up and restoring the "read" flag for both
sms and mms messages. Also, added a "call" method to the MmsSmsProvider
to return a boolean when a restore is in progress.

Test: updated the unit tests to test the new fields. Manually tested
a restore file to verify an mms with image was restored and imported
by bugle correctly.

Bug: 36367212
Bug: 36366196
Change-Id: If78bbe60a922aeb4373d45f9c82fa434806b13d9