Adding subscription group owner.
Define owner of a subscription group. Such that when subscriptions are
being added into or removed from a group, as long as callingPackage is
the group owner, it can skip carrier permission check of subscriptions
in the group.

This is necessary when a subscription is being added into a
group that contains inactive subscription, which mean its carrier
privilege can't be verified.

Bug: 131854492
Test: cts, unittest, TelephonyManagerTestApp
Change-Id: I365cca6302ca1ad830751bbfeb4d8d6d76d4e32c
Merged-In: I365cca6302ca1ad830751bbfeb4d8d6d76d4e32c