Remove APN db reading loophole
Remove the loophole for reading certain APN fields without privileges.
This was originally meant for MMS apps but they have a dedicated API for
what they're doing now.

Note that this change is not for backporting. It's intended to go into
API 30 (R). A separate change will be made for 29-and-below.

Bug: 124107808
Test: CTS
AOSP: Infeasible, security fix.
Change-Id: Ia7be0ecdfca5e7adf6163dc015a413a98f1b9287