Fix telephony backup creating diff when there is none.
The telephony provider recreates the message files for each
backup. This has the unfortunate side effect of making the
file modified time the time of the backup, which means the
diffing algorithm used for backups things the files have
changed.

To avoid this happening when there's no data change this CL
changes the backup agent to use the date from the most recent
message in a file as the file mtime, which should maintain
consistency between backup generations.

Bug: 140995339
Test: atest TelephonyProviderTests
Change-Id: I28753f2d92ef842d7ab4612913109a82d64d4d0a