Fix the default APN restoration thread always hold the TelephonyProvider lock.
The APN restoration with service acquired two locks and waited for binding service which caused other thread cannot access TelephonyProvider.
Since the symptom cannot be reproduced and there's no more information to find out the reason why it took so much time to bind service.
Add more log and release the lock to check if the binding is dead and make sure the TelephonyProvider won't be continuously locked.

Bug: 146932662
Test: atest TelephonyProviderTest
Change-Id: Id47bcb894b16844602d890e2f9b371c86dd9ccca