TelephonyBackupAgent: Catch only unchecked exceptions.
http://ag/12349208 introduced a catch block to be able to
continue work after non-critical exceptions such as the
SQLiteException encountered in bug 160801497.

Specifically, the catch block introducded by the above
CL catches any kind of Exception; because the try block
right now does not throw any checked exceptions, this
is equivalent to only catching RuntimeException. Catching
Exception has the downside that if a future CL changed
the code in the try block to throw some checked exception,
this would not become apparent because the code would still
compile.

This CL changes the catch block to catch RuntimeException,
which is arguably still a little wider than it needs to be
(SQLiteException might be better), but is guaranteed to not
introduce any behavior change relative to before this CL,
and is safer for future refatorings.

Test: Treehugger only; if this code compiles, it proves that
      no checked exception is thrown here.
Bug: 160801497

Change-Id: I066f07c320e02cf01903792b86a5877a9436c2b9