TelephonyBackupAgentTest: Clean up after ourselves.
Commit d99a1690d10badee0fc97e756902e98c2e3f0a79 in http://ag/12349208
introduced a @VisibleForTesting method in TelephonyBackupAgent that
swapped out mTelephonyBackupAgent's mSmsProviderQuery. This happened
to be safe because mTelephonyBackupAgent is re-instantiated in setUp(),
and this code was at the end of the test method. Still, it seems more
idiomatic and safer against future refactorings to restore the original
state at the end of the test. This CL does that.

Bug: 160801497
Test: atest TelephonyProviderTests:com.android.providers.telephony.TelephonyBackupAgentTest

Change-Id: I27d622773065e3e1ed8df46340a03ad79236085e