Parsing mcc and mnc as integers when null does not work
When overlay parsing for mcc and mnc was added to make sure that
the RRO apns got the correct format we did not count
on it being null. This causes a NumberformatException.

Since the format parsing was only done/needed for RROs this
was not seen until only carrier id was used for overlays.

Test: Add an APN in a RRO with only carrier id and no mcc/mnc data
Bug: 217871446
Change-Id: I3269a201fe902e96675ec5b73ef948cedc5b17ee