carrier id APNs duplicates in restore
When doing reset to default in APN Settings, APNs defined with only
carrier id are kept in the database and an extra APN is added.
For every restore there is an additional APN added which means that
the DB increase every time a user choose "reset to default".

This happens due to carrier id only APNs are added with NULL value
for mcc, mnc and numeric, while their default value is an empty string.
- Changing to set default empty string for numeric, mcc and mnc will
solve this problem.

The database query when deleting preferred APNs is also wrong since
it only deletes preferred APNs with a numeric value, it does not
take into account that a carrier id APN can be default.
- Adding carrier id in the where clause will do the trick here.

Test: Add a carrier id APN, insert its sim and try to restore
Bug: 221220163
Change-Id: I529bf8b5d2877b0c27ec682041e10fcf6fb1bc2d