Disallow interact with non-existent subscription
Before the change, messenge app can insert an entry that's associated
with a non-existent sub Id. However, later telephony doesn't allow the
app to query such entry(because subscriptionManager is unaware of such
non-existent sub Id), hence it's inconsistent behavior.

After the change, interactions with non-existent sub is disallowed,
except for -1 which is used for restoring bakcup.

Bug: 294125411
Test: Manually sending/receiving SMS/MMS,
    atest com.android.providers.telephony.SmsProviderTest,
    atest com.android.providers.telephony.MmsProviderTest,
    atest com.android.providers.telephony.ProviderUtilTest,
    atest CtsTelephonyProviderTestCases,
    atest CtsTelephonyTestCases
Change-Id: I70efd135f8932d84bb139632614a0533a6ca8ee6