Only load APNs once at first boot
The APN database is initited and populated at first boot. Currently this
is done twice as the database reference is grabbed after the checksum
shared pref is initiated. Since this large list is read twice in onCreate
it may lead to an ANR on low-end devices.

More details:
1. On first boot the database and the checksum shared pref does not exist
2. in updateApnDb() we first check if the checksum match, since it is not
set at first boot we continue.
3. Since the database does not exist, getWritableDatabase() will trigger
onCreate() of the DatabaseHelper see:
SQLiteOpenHelper#onCreate(android.database.sqlite.SQLiteDatabase)
4. onCreate() will create the database, do the initial import of APNs and set
the checksum shared pref.
5. We will then move on and read the apn list a second time here when
initDatabaseWithDatabaseHelper() call initDatabase() wihtout checking if it
is needed. In the end of initDatabase() the shared pref is set a second time.

Conclusion:
If we move the creation of database to happen before we check if the APN
list or the RROs have beeen updated we avoid that the list is read and
parsed twice.

Call sequence during first boot before this change:
TelephonyProvider.onCreate()
`-> updateApnDb()
`-> needApnDbUpdate() (true, on first boot as shared pref is not set)
`-> getWritableDatabase()
    `-> DatabaseHelper.onCreate()
        `-> initDatabase()
.....
`-> initDatabaseWithDatabaseHelper()
.....

Call sequence during first boot with this change:
TelephonyProvider.onCreate()
`-> updateApnDb()
`-> getWritableDatabase()
    `-> DatabaseHelper.onCreate()
        `-> initDatabase()
`-> needApnDbUpdate() (false, as shared pref is set by initDatabase())

Test: manual (get internet data on one carrier sim card and hot-swap to another carrier sim)
Test: atest TelephonyProviderTests
Bug: 303098050
Change-Id: Id78b7d975e344506dc8e2c0d6e1e9244322cac6a