Avoid deadlock when calling SubscriptionManager API
There were two locks that worked in opposite directions. The main
thread locked A and then waits B. The binder thread locked B and
then waits A. this is a deadlock.

Case: SubscriptionDatabaseManager#insertSubscriptionInfo held the
ReadWrite lock and waited for the TelephonyProvider to release the
its own lock. TelephonyProvider held its own lock and looked up
SubscriptionManager#isActiveSubscription and waited for the
ReadWrite lock to be released.

To avoid deadlock, The circular lock sequence must be broken.

Since SubscriptionDatabaseManager handles the TelephonyProvider's
database, we believed that the best design would be to first place
a parent lock on SubscriptionDatabaseManager and then lock
TelephonyProvider later, so modified it so that isActiveSubscription
can be queried without TelephonyProvider being locked.

Bug: 315417504
Test: atest TelephonyProviderTest
Change-Id: I507dfbca6a6d1b8369741e5481f71357fc5ce682