Telephony Provider version for HSUM
Provides an alternative TelephonyProvider version that is applicable to
the Headless System User Mode (HSUM), in which user 0 isn't a human
user. In this case, each user should have their own Messaging database,
since only the MainUser's TelephonyProvider will be supported.

The only difference between this and the regular TelephonyProvider is
the removal of some of the singleUser attributes.

Bug: 331960280
Bug: 347341495
Flag: EXEMPT xml change; also won't affect existing products

Test: Manually verified that the provider runs as a separate process in
      the MainUser and that it no longer has the singleUser attribute:
         adb shell dumpsys activity | grep -7 ".MmsSmsProvider"
Change-Id: Ic0323b5735acb39bfeea4c88b859f4da1c525474